requirements.txt
apipkg==1.5
asgiref==3.3.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
coverage==5.3
cryptography==3.2.1
dj-database-url==0.5.0
Django==3.1.3
django-crontab==0.7.1
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==3.0.0
django-polymorphic-tree==1.5.1
django-tag-parser==3.2
execnet==1.7.1
future==0.18.2
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
iniconfig==1.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.22
oauthlib==3.1.0
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
protobuf==3.14.0
psycopg2==2.8.6
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.1
pydotplus==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-forked==1.3.0
pytest-xdist==2.1.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
sentry-sdk==0.19.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
waitress==1.4.4
WebOb==1.8.6
WebTest==2.0.35
whitenoise==5.2.0
wrapt==1.12.1
